[core] add separate `neighbor.hpp`, `child.hpp`, and `router.hpp` (#9376)
This commit adds separate files for `Neighbor`, `Child`, and `Router`
classes.
