+++ src/core/BUILD.gn
@@ -604,6 +604,8 @@ openthread_core_files = [
   "thread/announce_sender.hpp",
   "thread/anycast_locator.cpp",
   "thread/anycast_locator.hpp",
+  "thread/child.cpp",
+  "thread/child.hpp",
   "thread/child_mask.hpp",
   "thread/child_supervision.cpp",
   "thread/child_supervision.hpp",
@@ -646,6 +648,8 @@ openthread_core_files = [
   "thread/mlr_manager.cpp",
   "thread/mlr_manager.hpp",
   "thread/mlr_types.hpp",
+  "thread/neighbor.cpp",
+  "thread/neighbor.hpp",
   "thread/neighbor_table.cpp",
   "thread/neighbor_table.hpp",
   "thread/network_data.cpp",
@@ -674,6 +678,8 @@ openthread_core_files = [
   "thread/panid_query_server.hpp",
   "thread/radio_selector.cpp",
   "thread/radio_selector.hpp",
+  "thread/router.cpp",
+  "thread/router.hpp",
   "thread/router_table.cpp",
   "thread/router_table.hpp",
   "thread/src_match_controller.cpp",
@@ -685,8 +691,6 @@ openthread_core_files = [
   "thread/time_sync_service.hpp",
   "thread/tmf.cpp",
   "thread/tmf.hpp",
-  "thread/topology.cpp",
-  "thread/topology.hpp",
   "thread/uri_paths.cpp",
   "thread/uri_paths.hpp",
   "thread/version.hpp",
