+++ src/core/thread/child.cpp
@@ -28,10 +28,10 @@
 
 /**
  * @file
- *   This file includes definitions for maintaining Thread network topologies.
+ *   This file includes definitions for a Thread `Child`.
  */
 
-#include "topology.hpp"
+#include "child.hpp"
 
 #include "common/array.hpp"
 #include "common/code_utils.hpp"
@@ -41,222 +41,6 @@
 #include "common/num_utils.hpp"
 
 namespace ot {
-
-void Neighbor::SetState(State aState)
-{
-    VerifyOrExit(mState != aState);
-    mState = static_cast<uint8_t>(aState);
-
-#if OPENTHREAD_CONFIG_UPTIME_ENABLE
-    if (mState == kStateValid)
-    {
-        mConnectionStart = Uptime::MsecToSec(Get<Uptime>().GetUptime());
-    }
-#endif
-
-exit:
-    return;
-}
-
-#if OPENTHREAD_CONFIG_UPTIME_ENABLE
-uint32_t Neighbor::GetConnectionTime(void) const
-{
-    return IsStateValid() ? Uptime::MsecToSec(Get<Uptime>().GetUptime()) - mConnectionStart : 0;
-}
-#endif
-
-bool Neighbor::AddressMatcher::Matches(const Neighbor &aNeighbor) const
-{
-    bool matches = false;
-
-    VerifyOrExit(aNeighbor.MatchesFilter(mStateFilter));
-
-    if (mShortAddress != Mac::kShortAddrInvalid)
-    {
-        VerifyOrExit(mShortAddress == aNeighbor.GetRloc16());
-    }
-
-    if (mExtAddress != nullptr)
-    {
-        VerifyOrExit(*mExtAddress == aNeighbor.GetExtAddress());
-    }
-
-    matches = true;
-
-exit:
-    return matches;
-}
-
-void Neighbor::Info::SetFrom(const Neighbor &aNeighbor)
-{
-    Clear();
-
-    mExtAddress       = aNeighbor.GetExtAddress();
-    mAge              = Time::MsecToSec(TimerMilli::GetNow() - aNeighbor.GetLastHeard());
-    mRloc16           = aNeighbor.GetRloc16();
-    mLinkFrameCounter = aNeighbor.GetLinkFrameCounters().GetMaximum();
-    mMleFrameCounter  = aNeighbor.GetMleFrameCounter();
-    mLinkQualityIn    = aNeighbor.GetLinkQualityIn();
-    mAverageRssi      = aNeighbor.GetLinkInfo().GetAverageRss();
-    mLastRssi         = aNeighbor.GetLinkInfo().GetLastRss();
-    mLinkMargin       = aNeighbor.GetLinkInfo().GetLinkMargin();
-    mFrameErrorRate   = aNeighbor.GetLinkInfo().GetFrameErrorRate();
-    mMessageErrorRate = aNeighbor.GetLinkInfo().GetMessageErrorRate();
-    mRxOnWhenIdle     = aNeighbor.IsRxOnWhenIdle();
-    mFullThreadDevice = aNeighbor.IsFullThreadDevice();
-    mFullNetworkData  = (aNeighbor.GetNetworkDataType() == NetworkData::kFullSet);
-    mVersion          = aNeighbor.GetVersion();
-#if OPENTHREAD_CONFIG_UPTIME_ENABLE
-    mConnectionTime = aNeighbor.GetConnectionTime();
-#endif
-}
-
-void Neighbor::Init(Instance &aInstance)
-{
-    InstanceLocatorInit::Init(aInstance);
-    mLinkInfo.Init(aInstance);
-    SetState(kStateInvalid);
-}
-
-bool Neighbor::IsStateValidOrAttaching(void) const
-{
-    bool rval = false;
-
-    switch (GetState())
-    {
-    case kStateInvalid:
-    case kStateParentRequest:
-    case kStateParentResponse:
-        break;
-
-    case kStateRestored:
-    case kStateChildIdRequest:
-    case kStateLinkRequest:
-    case kStateChildUpdateRequest:
-    case kStateValid:
-        rval = true;
-        break;
-    }
-
-    return rval;
-}
-
-bool Neighbor::MatchesFilter(StateFilter aFilter) const
-{
-    bool matches = false;
-
-    switch (aFilter)
-    {
-    case kInStateValid:
-        matches = IsStateValid();
-        break;
-
-    case kInStateValidOrRestoring:
-        matches = IsStateValidOrRestoring();
-        break;
-
-    case kInStateChildIdRequest:
-        matches = IsStateChildIdRequest();
-        break;
-
-    case kInStateValidOrAttaching:
-        matches = IsStateValidOrAttaching();
-        break;
-
-    case kInStateInvalid:
-        matches = IsStateInvalid();
-        break;
-
-    case kInStateAnyExceptInvalid:
-        matches = !IsStateInvalid();
-        break;
-
-    case kInStateAnyExceptValidOrRestoring:
-        matches = !IsStateValidOrRestoring();
-        break;
-
-    case kInStateAny:
-        matches = true;
-        break;
-    }
-
-    return matches;
-}
-
-#if OPENTHREAD_CONFIG_MULTI_RADIO
-void Neighbor::SetLastRxFragmentTag(uint16_t aTag)
-{
-    mLastRxFragmentTag     = (aTag == 0) ? 0xffff : aTag;
-    mLastRxFragmentTagTime = TimerMilli::GetNow();
-}
-
-bool Neighbor::IsLastRxFragmentTagSet(void) const
-{
-    return (mLastRxFragmentTag != 0) && (TimerMilli::GetNow() <= mLastRxFragmentTagTime + kLastRxFragmentTagTimeout);
-}
-#endif
-
-#if OPENTHREAD_CONFIG_MLE_LINK_METRICS_SUBJECT_ENABLE
-void Neighbor::AggregateLinkMetrics(uint8_t aSeriesId, uint8_t aFrameType, uint8_t aLqi, int8_t aRss)
-{
-    for (LinkMetrics::SeriesInfo &entry : mLinkMetricsSeriesInfoList)
-    {
-        if (aSeriesId == 0 || aSeriesId == entry.GetSeriesId())
-        {
-            entry.AggregateLinkMetrics(aFrameType, aLqi, aRss);
-        }
-    }
-}
-
-LinkMetrics::SeriesInfo *Neighbor::GetForwardTrackingSeriesInfo(const uint8_t &aSeriesId)
-{
-    return mLinkMetricsSeriesInfoList.FindMatching(aSeriesId);
-}
-
-void Neighbor::AddForwardTrackingSeriesInfo(LinkMetrics::SeriesInfo &aSeriesInfo)
-{
-    mLinkMetricsSeriesInfoList.Push(aSeriesInfo);
-}
-
-LinkMetrics::SeriesInfo *Neighbor::RemoveForwardTrackingSeriesInfo(const uint8_t &aSeriesId)
-{
-    return mLinkMetricsSeriesInfoList.RemoveMatching(aSeriesId);
-}
-
-void Neighbor::RemoveAllForwardTrackingSeriesInfo(void)
-{
-    while (!mLinkMetricsSeriesInfoList.IsEmpty())
-    {
-        LinkMetrics::SeriesInfo *seriesInfo = mLinkMetricsSeriesInfoList.Pop();
-        Get<LinkMetrics::Subject>().Free(*seriesInfo);
-    }
-}
-#endif // OPENTHREAD_CONFIG_MLE_LINK_METRICS_SUBJECT_ENABLE
-
-const char *Neighbor::StateToString(State aState)
-{
-    static const char *const kStateStrings[] = {
-        "Invalid",        // (0) kStateInvalid
-        "Restored",       // (1) kStateRestored
-        "ParentReq",      // (2) kStateParentRequest
-        "ParentRes",      // (3) kStateParentResponse
-        "ChildIdReq",     // (4) kStateChildIdRequest
-        "LinkReq",        // (5) kStateLinkRequest
-        "ChildUpdateReq", // (6) kStateChildUpdateRequest
-        "Valid",          // (7) kStateValid
-    };
-
-    static_assert(0 == kStateInvalid, "kStateInvalid value is incorrect");
-    static_assert(1 == kStateRestored, "kStateRestored value is incorrect");
-    static_assert(2 == kStateParentRequest, "kStateParentRequest value is incorrect");
-    static_assert(3 == kStateParentResponse, "kStateParentResponse value is incorrect");
-    static_assert(4 == kStateChildIdRequest, "kStateChildIdRequest value is incorrect");
-    static_assert(5 == kStateLinkRequest, "kStateLinkRequest value is incorrect");
-    static_assert(6 == kStateChildUpdateRequest, "kStateChildUpdateRequest value is incorrect");
-    static_assert(7 == kStateValid, "kStateValid value is incorrect");
-
-    return kStateStrings[aState];
-}
 
 #if OPENTHREAD_FTD
 
@@ -532,80 +316,5 @@ void Child::SetAddressMlrState(const Ip6::Address &aAddress, MlrState aState)
 #endif // OPENTHREAD_CONFIG_TMF_PROXY_MLR_ENABLE
 
 #endif // OPENTHREAD_FTD
-
-void Router::Info::SetFrom(const Router &aRouter)
-{
-    Clear();
-    mRloc16          = aRouter.GetRloc16();
-    mRouterId        = Mle::RouterIdFromRloc16(mRloc16);
-    mExtAddress      = aRouter.GetExtAddress();
-    mAllocated       = true;
-    mNextHop         = aRouter.GetNextHop();
-    mLinkEstablished = aRouter.IsStateValid();
-    mPathCost        = aRouter.GetCost();
-    mLinkQualityIn   = aRouter.GetLinkQualityIn();
-    mLinkQualityOut  = aRouter.GetLinkQualityOut();
-    mAge             = static_cast<uint8_t>(Time::MsecToSec(TimerMilli::GetNow() - aRouter.GetLastHeard()));
-    mVersion         = ClampToUint8(aRouter.GetVersion());
-}
-
-void Router::Info::SetFrom(const Parent &aParent)
-{
-    SetFrom(static_cast<const Router &>(aParent));
-#if OPENTHREAD_CONFIG_MAC_CSL_RECEIVER_ENABLE
-    mCslClockAccuracy = aParent.GetCslAccuracy().GetClockAccuracy();
-    mCslUncertainty   = aParent.GetCslAccuracy().GetUncertainty();
-#endif
-}
-
-void Router::Clear(void)
-{
-    Instance &instance = GetInstance();
-
-    memset(reinterpret_cast<void *>(this), 0, sizeof(Router));
-    Init(instance);
-}
-
-LinkQuality Router::GetTwoWayLinkQuality(void) const { return Min(GetLinkQualityIn(), GetLinkQualityOut()); }
-
-void Router::SetFrom(const Parent &aParent)
-{
-    // We use an intermediate pointer to copy `aParent` to silence
-    // code checkers warning about object slicing (assigning a
-    // sub-class to base class instance).
-
-    const Router *parentAsRouter = &aParent;
-
-    *this = *parentAsRouter;
-}
-
-void Parent::Clear(void)
-{
-    Instance &instance = GetInstance();
-
-    memset(reinterpret_cast<void *>(this), 0, sizeof(Parent));
-    Init(instance);
-}
-
-bool Router::SetNextHopAndCost(uint8_t aNextHop, uint8_t aCost)
-{
-    bool changed = false;
-
-    if (mNextHop != aNextHop)
-    {
-        mNextHop = aNextHop;
-        changed  = true;
-    }
-
-    if (mCost != aCost)
-    {
-        mCost   = aCost;
-        changed = true;
-    }
-
-    return changed;
-}
-
-bool Router::SetNextHopToInvalid(void) { return SetNextHopAndCost(Mle::kInvalidRouterId, 0); }
 
 } // namespace ot
