+++ src/core/thread/child_supervision.hpp
@@ -46,12 +46,11 @@
 #include "common/time_ticker.hpp"
 #include "common/timer.hpp"
 #include "mac/mac_types.hpp"
-#include "thread/topology.hpp"
+#include "thread/child.hpp"
 
 namespace ot {
 
 class ThreadNetif;
-class Child;
 
 /**
  *
