+++ src/core/thread/indirect_sender.cpp
@@ -39,9 +39,9 @@
 #include "common/instance.hpp"
 #include "common/locator_getters.hpp"
 #include "common/message.hpp"
+#include "thread/child.hpp"
 #include "thread/mesh_forwarder.hpp"
 #include "thread/mle_tlvs.hpp"
-#include "thread/topology.hpp"
 
 namespace ot {
