+++ src/core/thread/mesh_forwarder.hpp
@@ -50,10 +50,10 @@
 #include "mac/mac_frame.hpp"
 #include "net/ip6.hpp"
 #include "thread/address_resolver.hpp"
+#include "thread/child.hpp"
 #include "thread/indirect_sender.hpp"
 #include "thread/lowpan.hpp"
 #include "thread/network_data_leader.hpp"
-#include "thread/topology.hpp"
 
 namespace ot {
