+++ src/core/thread/mlr_manager.hpp
@@ -51,8 +51,8 @@
 #include "common/time_ticker.hpp"
 #include "common/timer.hpp"
 #include "net/netif.hpp"
+#include "thread/child.hpp"
 #include "thread/thread_tlvs.hpp"
-#include "thread/topology.hpp"
 
 namespace ot {
