+++ src/core/thread/router_table.hpp
@@ -44,8 +44,8 @@
 #include "mac/mac_types.hpp"
 #include "thread/mle_tlvs.hpp"
 #include "thread/mle_types.hpp"
+#include "thread/router.hpp"
 #include "thread/thread_tlvs.hpp"
-#include "thread/topology.hpp"
 
 namespace ot {
