+++ tests/unit/test_child.cpp
@@ -33,7 +33,7 @@
 #include "test_util.h"
 #include "common/code_utils.hpp"
 #include "common/instance.hpp"
-#include "thread/topology.hpp"
+#include "thread/child.hpp"
 
 namespace ot {
